+++ CHANGELOG.md
@@ -1,7 +1,28 @@
 Changelog
 =========
 
+2.0.0
+-----
+
+- [Brk] PSR12 support
+- [Brk] Namespace change.
+
 1.0.0
+-----
+- [Enh] being able to extend the definition for the user foreign keys
+
+0.1.2
+-----
+
+- [Bug] translations not always loaded
+
+
+0.1.1
+-----
+
+- [Bug] make `tecnocen\rmdb\Module:userId` readable.
+
+0.1.0
 -----
 
 - [Enh] migration to create RMDB tables